requirements.txt
django>=2.1.2,<=2.2
django-crispy-forms>=1.7.2,<=1.8
django-widget-tweaks>=1.4.3,<=1.5
